requirements.txt
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-WTF==0.14.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
WTForms-Components==0.10.4
WTForms-SQLAlchemy==0.1
